LayoutTests/platform/mac/css3/filters/custom/custom-filter-property-computed-style-expected.txt

Tests the computed style of the custom() function of the -webkit-filter property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Custom with vertex shader : custom(url(vertex.shader))
FAIL removeBaseURL(filterStyle.getPropertyValue('-webkit-filter')) should be custom(url(vertex.shader) none, 1 1 filter-box). Was none.
FAIL jsWrapperClass(filterRule) should be CSSValueList. Was CSSPrimitiveValue.
FAIL jsWrapperClass(filterRule.__proto__) should be CSSValueListPrototype. Was CSSPrimitiveValuePrototype.
FAIL jsWrapperClass(filterRule.constructor) should be CSSValueListConstructor. Was CSSPrimitiveValueConstructor.
FAIL filterRule.length should be 1 (of type number). Was undefined (of type undefined).
FAIL subRule.operationType should be 12. Threw exception TypeError: 'undefined' is not an object (evaluating 'subRule.operationType')
FAIL removeBaseURL(subRule.cssText) should be custom(url(vertex.shader) none, 1 1 filter-box). Threw exception TypeError: 'undefined' is not an object (evaluating 'subRule.cssText')

Custom with fragment shader : custom(none url(fragment.shader))
FAIL removeBaseURL(filterStyle.getPropertyValue('-webkit-filter')) should be custom(none url(fragment.shader), 1 1 filter-box). Was none.
FAIL jsWrapperClass(filterRule) should be CSSValueList. Was CSSPrimitiveValue.
FAIL jsWrapperClass(filterRule.__proto__) should be CSSValueListPrototype. Was CSSPrimitiveValuePrototype.
FAIL jsWrapperClass(filterRule.constructor) should be CSSValueListConstructor. Was CSSPrimitiveValueConstructor.
FAIL filterRule.length should be 1 (of type number). Was undefined (of type undefined).
FAIL subRule.operationType should be 12. Threw exception TypeError: 'undefined' is not an object (evaluating 'subRule.operationType')
FAIL removeBaseURL(subRule.cssText) should be custom(none url(fragment.shader), 1 1 filter-box). Threw exception TypeError: 'undefined' is not an object (evaluating 'subRule.cssText')

Custom with both shaders : custom(url(vertex.shader) url(fragment.shader))
FAIL removeBaseURL(filterStyle.getPropertyValue('-webkit-filter')) should be custom(url(vertex.shader) url(fragment.shader), 1 1 filter-box). Was none.
FAIL jsWrapperClass(filterRule) should be CSSValueList. Was CSSPrimitiveValue.
FAIL jsWrapperClass(filterRule.__proto__) should be CSSValueListPrototype. Was CSSPrimitiveValuePrototype.
FAIL jsWrapperClass(filterRule.constructor) should be CSSValueListConstructor. Was CSSPrimitiveValueConstructor.
FAIL filterRule.length should be 1 (of type number). Was undefined (of type undefined).
FAIL subRule.operationType should be 12. Threw exception TypeError: 'undefined' is not an object (evaluating 'subRule.operationType')
FAIL removeBaseURL(subRule.cssText) should be custom(url(vertex.shader) url(fragment.shader), 1 1 filter-box). Threw exception TypeError: 'undefined' is not an object (evaluating 'subRule.cssText')

Custom with mesh size : custom(url(vertex.shader), 10)
FAIL removeBaseURL(filterStyle.getPropertyValue('-webkit-filter')) should be custom(url(vertex.shader) none, 10 10 filter-box). Was none.
FAIL jsWrapperClass(filterRule) should be CSSValueList. Was CSSPrimitiveValue.
FAIL jsWrapperClass(filterRule.__proto__) should be CSSValueListPrototype. Was CSSPrimitiveValuePrototype.
FAIL jsWrapperClass(filterRule.constructor) should be CSSValueListConstructor. Was CSSPrimitiveValueConstructor.
FAIL filterRule.length should be 1 (of type number). Was undefined (of type undefined).
FAIL subRule.operationType should be 12. Threw exception TypeError: 'undefined' is not an object (evaluating 'subRule.operationType')
FAIL removeBaseURL(subRule.cssText) should be custom(url(vertex.shader) none, 10 10 filter-box). Threw exception TypeError: 'undefined' is not an object (evaluating 'subRule.cssText')

Custom with both mesh sizes : custom(none url(fragment.shader), 10 20)
FAIL removeBaseURL(filterStyle.getPropertyValue('-webkit-filter')) should be custom(none url(fragment.shader), 10 20 filter-box). Was none.
FAIL jsWrapperClass(filterRule) should be CSSValueList. Was CSSPrimitiveValue.
FAIL jsWrapperClass(filterRule.__proto__) should be CSSValueListPrototype. Was CSSPrimitiveValuePrototype.
FAIL jsWrapperClass(filterRule.constructor) should be CSSValueListConstructor. Was CSSPrimitiveValueConstructor.
FAIL filterRule.length should be 1 (of type number). Was undefined (of type undefined).
FAIL subRule.operationType should be 12. Threw exception TypeError: 'undefined' is not an object (evaluating 'subRule.operationType')
FAIL removeBaseURL(subRule.cssText) should be custom(none url(fragment.shader), 10 20 filter-box). Threw exception TypeError: 'undefined' is not an object (evaluating 'subRule.cssText')

Custom with detached mesh : custom(none url(fragment.shader), detached)
FAIL removeBaseURL(filterStyle.getPropertyValue('-webkit-filter')) should be custom(none url(fragment.shader), 1 1 filter-box detached). Was none.
FAIL jsWrapperClass(filterRule) should be CSSValueList. Was CSSPrimitiveValue.
FAIL jsWrapperClass(filterRule.__proto__) should be CSSValueListPrototype. Was CSSPrimitiveValuePrototype.
FAIL jsWrapperClass(filterRule.constructor) should be CSSValueListConstructor. Was CSSPrimitiveValueConstructor.
FAIL filterRule.length should be 1 (of type number). Was undefined (of type undefined).
FAIL subRule.operationType should be 12. Threw exception TypeError: 'undefined' is not an object (evaluating 'subRule.operationType')
FAIL removeBaseURL(subRule.cssText) should be custom(none url(fragment.shader), 1 1 filter-box detached). Threw exception TypeError: 'undefined' is not an object (evaluating 'subRule.cssText')

Custom with detached and mesh size : custom(none url(fragment.shader), 10 20 detached)
FAIL removeBaseURL(filterStyle.getPropertyValue('-webkit-filter')) should be custom(none url(fragment.shader), 10 20 filter-box detached). Was none.
FAIL jsWrapperClass(filterRule) should be CSSValueList. Was CSSPrimitiveValue.
FAIL jsWrapperClass(filterRule.__proto__) should be CSSValueListPrototype. Was CSSPrimitiveValuePrototype.
FAIL jsWrapperClass(filterRule.constructor) should be CSSValueListConstructor. Was CSSPrimitiveValueConstructor.
FAIL filterRule.length should be 1 (of type number). Was undefined (of type undefined).
FAIL subRule.operationType should be 12. Threw exception TypeError: 'undefined' is not an object (evaluating 'subRule.operationType')
FAIL removeBaseURL(subRule.cssText) should be custom(none url(fragment.shader), 10 20 filter-box detached). Threw exception TypeError: 'undefined' is not an object (evaluating 'subRule.cssText')

Custom with default filter-box : custom(none url(fragment.shader), filter-box)
FAIL removeBaseURL(filterStyle.getPropertyValue('-webkit-filter')) should be custom(none url(fragment.shader), 1 1 filter-box). Was none.
FAIL jsWrapperClass(filterRule) should be CSSValueList. Was CSSPrimitiveValue.
FAIL jsWrapperClass(filterRule.__proto__) should be CSSValueListPrototype. Was CSSPrimitiveValuePrototype.
FAIL jsWrapperClass(filterRule.constructor) should be CSSValueListConstructor. Was CSSPrimitiveValueConstructor.
FAIL filterRule.length should be 1 (of type number). Was undefined (of type undefined).
FAIL subRule.operationType should be 12. Threw exception TypeError: 'undefined' is not an object (evaluating 'subRule.operationType')
FAIL removeBaseURL(subRule.cssText) should be custom(none url(fragment.shader), 1 1 filter-box). Threw exception TypeError: 'undefined' is not an object (evaluating 'subRule.cssText')

Custom with content-box : custom(none url(fragment.shader), content-box)
FAIL removeBaseURL(filterStyle.getPropertyValue('-webkit-filter')) should be custom(none url(fragment.shader), 1 1 content-box). Was none.
FAIL jsWrapperClass(filterRule) should be CSSValueList. Was CSSPrimitiveValue.
FAIL jsWrapperClass(filterRule.__proto__) should be CSSValueListPrototype. Was CSSPrimitiveValuePrototype.
FAIL jsWrapperClass(filterRule.constructor) should be CSSValueListConstructor. Was CSSPrimitiveValueConstructor.
FAIL filterRule.length should be 1 (of type number). Was undefined (of type undefined).
FAIL subRule.operationType should be 12. Threw exception TypeError: 'undefined' is not an object (evaluating 'subRule.operationType')
FAIL removeBaseURL(subRule.cssText) should be custom(none url(fragment.shader), 1 1 content-box). Threw exception TypeError: 'undefined' is not an object (evaluating 'subRule.cssText')

Custom with border-box : custom(none url(fragment.shader), border-box)
FAIL removeBaseURL(filterStyle.getPropertyValue('-webkit-filter')) should be custom(none url(fragment.shader), 1 1 border-box). Was none.
FAIL jsWrapperClass(filterRule) should be CSSValueList. Was CSSPrimitiveValue.
FAIL jsWrapperClass(filterRule.__proto__) should be CSSValueListPrototype. Was CSSPrimitiveValuePrototype.
FAIL jsWrapperClass(filterRule.constructor) should be CSSValueListConstructor. Was CSSPrimitiveValueConstructor.
FAIL filterRule.length should be 1 (of type number). Was undefined (of type undefined).
FAIL subRule.operationType should be 12. Threw exception TypeError: 'undefined' is not an object (evaluating 'subRule.operationType')
FAIL removeBaseURL(subRule.cssText) should be custom(none url(fragment.shader), 1 1 border-box). Threw exception TypeError: 'undefined' is not an object (evaluating 'subRule.cssText')

Custom with padding-box : custom(none url(fragment.shader), padding-box)
FAIL removeBaseURL(filterStyle.getPropertyValue('-webkit-filter')) should be custom(none url(fragment.shader), 1 1 padding-box). Was none.
FAIL jsWrapperClass(filterRule) should be CSSValueList. Was CSSPrimitiveValue.
FAIL jsWrapperClass(filterRule.__proto__) should be CSSValueListPrototype. Was CSSPrimitiveValuePrototype.
FAIL jsWrapperClass(filterRule.constructor) should be CSSValueListConstructor. Was CSSPrimitiveValueConstructor.
FAIL filterRule.length should be 1 (of type number). Was undefined (of type undefined).
FAIL subRule.operationType should be 12. Threw exception TypeError: 'undefined' is not an object (evaluating 'subRule.operationType')
FAIL removeBaseURL(subRule.cssText) should be custom(none url(fragment.shader), 1 1 padding-box). Threw exception TypeError: 'undefined' is not an object (evaluating 'subRule.cssText')

Custom with mesh-size and padding-box : custom(none url(fragment.shader), 10 padding-box)
FAIL removeBaseURL(filterStyle.getPropertyValue('-webkit-filter')) should be custom(none url(fragment.shader), 10 10 padding-box). Was none.
FAIL jsWrapperClass(filterRule) should be CSSValueList. Was CSSPrimitiveValue.
FAIL jsWrapperClass(filterRule.__proto__) should be CSSValueListPrototype. Was CSSPrimitiveValuePrototype.
FAIL jsWrapperClass(filterRule.constructor) should be CSSValueListConstructor. Was CSSPrimitiveValueConstructor.
FAIL filterRule.length should be 1 (of type number). Was undefined (of type undefined).
FAIL subRule.operationType should be 12. Threw exception TypeError: 'undefined' is not an object (evaluating 'subRule.operationType')
FAIL removeBaseURL(subRule.cssText) should be custom(none url(fragment.shader), 10 10 padding-box). Threw exception TypeError: 'undefined' is not an object (evaluating 'subRule.cssText')

Custom with mesh-size and padding-box : custom(none url(fragment.shader), padding-box detached)
FAIL removeBaseURL(filterStyle.getPropertyValue('-webkit-filter')) should be custom(none url(fragment.shader), 1 1 padding-box detached). Was none.
FAIL jsWrapperClass(filterRule) should be CSSValueList. Was CSSPrimitiveValue.
FAIL jsWrapperClass(filterRule.__proto__) should be CSSValueListPrototype. Was CSSPrimitiveValuePrototype.
FAIL jsWrapperClass(filterRule.constructor) should be CSSValueListConstructor. Was CSSPrimitiveValueConstructor.
FAIL filterRule.length should be 1 (of type number). Was undefined (of type undefined).
FAIL subRule.operationType should be 12. Threw exception TypeError: 'undefined' is not an object (evaluating 'subRule.operationType')
FAIL removeBaseURL(subRule.cssText) should be custom(none url(fragment.shader), 1 1 padding-box detached). Threw exception TypeError: 'undefined' is not an object (evaluating 'subRule.cssText')

Custom with both mesh-sizes and padding-box : custom(none url(fragment.shader), 10 20 padding-box)
FAIL removeBaseURL(filterStyle.getPropertyValue('-webkit-filter')) should be custom(none url(fragment.shader), 10 20 padding-box). Was none.
FAIL jsWrapperClass(filterRule) should be CSSValueList. Was CSSPrimitiveValue.
FAIL jsWrapperClass(filterRule.__proto__) should be CSSValueListPrototype. Was CSSPrimitiveValuePrototype.
FAIL jsWrapperClass(filterRule.constructor) should be CSSValueListConstructor. Was CSSPrimitiveValueConstructor.
FAIL filterRule.length should be 1 (of type number). Was undefined (of type undefined).
FAIL subRule.operationType should be 12. Threw exception TypeError: 'undefined' is not an object (evaluating 'subRule.operationType')
FAIL removeBaseURL(subRule.cssText) should be custom(none url(fragment.shader), 10 20 padding-box). Threw exception TypeError: 'undefined' is not an object (evaluating 'subRule.cssText')

Custom with both mesh-sizes and padding-box and detached : custom(none url(fragment.shader), 10 20 padding-box detached)
FAIL removeBaseURL(filterStyle.getPropertyValue('-webkit-filter')) should be custom(none url(fragment.shader), 10 20 padding-box detached). Was none.
FAIL jsWrapperClass(filterRule) should be CSSValueList. Was CSSPrimitiveValue.
FAIL jsWrapperClass(filterRule.__proto__) should be CSSValueListPrototype. Was CSSPrimitiveValuePrototype.
FAIL jsWrapperClass(filterRule.constructor) should be CSSValueListConstructor. Was CSSPrimitiveValueConstructor.
FAIL filterRule.length should be 1 (of type number). Was undefined (of type undefined).
FAIL subRule.operationType should be 12. Threw exception TypeError: 'undefined' is not an object (evaluating 'subRule.operationType')
FAIL removeBaseURL(subRule.cssText) should be custom(none url(fragment.shader), 10 20 padding-box detached). Threw exception TypeError: 'undefined' is not an object (evaluating 'subRule.cssText')

Custom with padding-box and detached : custom(none url(fragment.shader), padding-box detached)
FAIL removeBaseURL(filterStyle.getPropertyValue('-webkit-filter')) should be custom(none url(fragment.shader), 1 1 padding-box detached). Was none.
FAIL jsWrapperClass(filterRule) should be CSSValueList. Was CSSPrimitiveValue.
FAIL jsWrapperClass(filterRule.__proto__) should be CSSValueListPrototype. Was CSSPrimitiveValuePrototype.
FAIL jsWrapperClass(filterRule.constructor) should be CSSValueListConstructor. Was CSSPrimitiveValueConstructor.
FAIL filterRule.length should be 1 (of type number). Was undefined (of type undefined).
FAIL subRule.operationType should be 12. Threw exception TypeError: 'undefined' is not an object (evaluating 'subRule.operationType')
FAIL removeBaseURL(subRule.cssText) should be custom(none url(fragment.shader), 1 1 padding-box detached). Threw exception TypeError: 'undefined' is not an object (evaluating 'subRule.cssText')

Custom with integer parameters : custom(none url(fragment.shader), p1 1, p2 2 3, p3 3 4 5, p4 4 5 6 7)
FAIL removeBaseURL(filterStyle.getPropertyValue('-webkit-filter')) should be custom(none url(fragment.shader), 1 1 filter-box, p1 1, p2 2 3, p3 3 4 5, p4 4 5 6 7). Was none.
FAIL jsWrapperClass(filterRule) should be CSSValueList. Was CSSPrimitiveValue.
FAIL jsWrapperClass(filterRule.__proto__) should be CSSValueListPrototype. Was CSSPrimitiveValuePrototype.
FAIL jsWrapperClass(filterRule.constructor) should be CSSValueListConstructor. Was CSSPrimitiveValueConstructor.
FAIL filterRule.length should be 1 (of type number). Was undefined (of type undefined).
FAIL subRule.operationType should be 12. Threw exception TypeError: 'undefined' is not an object (evaluating 'subRule.operationType')
FAIL removeBaseURL(subRule.cssText) should be custom(none url(fragment.shader), 1 1 filter-box, p1 1, p2 2 3, p3 3 4 5, p4 4 5 6 7). Threw exception TypeError: 'undefined' is not an object (evaluating 'subRule.cssText')

Custom with float parameters : custom(none url(fragment.shader), p1 1.1, p2 2.2 3.3, p3 3.1 4.1 5.1, p4 4.1 5.2 6.3 7.4)
FAIL removeBaseURL(filterStyle.getPropertyValue('-webkit-filter')) should be custom(none url(fragment.shader), 1 1 filter-box, p1 1.1, p2 2.2 3.3, p3 3.1 4.1 5.1, p4 4.1 5.2 6.3 7.4). Was none.
FAIL jsWrapperClass(filterRule) should be CSSValueList. Was CSSPrimitiveValue.
FAIL jsWrapperClass(filterRule.__proto__) should be CSSValueListPrototype. Was CSSPrimitiveValuePrototype.
FAIL jsWrapperClass(filterRule.constructor) should be CSSValueListConstructor. Was CSSPrimitiveValueConstructor.
FAIL filterRule.length should be 1 (of type number). Was undefined (of type undefined).
FAIL subRule.operationType should be 12. Threw exception TypeError: 'undefined' is not an object (evaluating 'subRule.operationType')
FAIL removeBaseURL(subRule.cssText) should be custom(none url(fragment.shader), 1 1 filter-box, p1 1.1, p2 2.2 3.3, p3 3.1 4.1 5.1, p4 4.1 5.2 6.3 7.4). Threw exception TypeError: 'undefined' is not an object (evaluating 'subRule.cssText')

Custom with mesh size and number parameters : custom(none url(fragment.shader), 10 20 filter-box, p1 1, p2 2 3, p3 3 4 5, p4 4 5 6 7)
FAIL removeBaseURL(filterStyle.getPropertyValue('-webkit-filter')) should be custom(none url(fragment.shader), 10 20 filter-box, p1 1, p2 2 3, p3 3 4 5, p4 4 5 6 7). Was none.
FAIL jsWrapperClass(filterRule) should be CSSValueList. Was CSSPrimitiveValue.
FAIL jsWrapperClass(filterRule.__proto__) should be CSSValueListPrototype. Was CSSPrimitiveValuePrototype.
FAIL jsWrapperClass(filterRule.constructor) should be CSSValueListConstructor. Was CSSPrimitiveValueConstructor.
FAIL filterRule.length should be 1 (of type number). Was undefined (of type undefined).
FAIL subRule.operationType should be 12. Threw exception TypeError: 'undefined' is not an object (evaluating 'subRule.operationType')
FAIL removeBaseURL(subRule.cssText) should be custom(none url(fragment.shader), 10 20 filter-box, p1 1, p2 2 3, p3 3 4 5, p4 4 5 6 7). Threw exception TypeError: 'undefined' is not an object (evaluating 'subRule.cssText')

Multiple with fragment shader : grayscale() custom(none url(fragment.shader)) sepia()
FAIL removeBaseURL(filterStyle.getPropertyValue('-webkit-filter')) should be grayscale(1) custom(none url(fragment.shader), 1 1 filter-box) sepia(1). Was none.
FAIL jsWrapperClass(filterRule) should be CSSValueList. Was CSSPrimitiveValue.
FAIL jsWrapperClass(filterRule.__proto__) should be CSSValueListPrototype. Was CSSPrimitiveValuePrototype.
FAIL jsWrapperClass(filterRule.constructor) should be CSSValueListConstructor. Was CSSPrimitiveValueConstructor.
FAIL filterRule.length should be 3 (of type number). Was undefined (of type undefined).
FAIL subRule.operationType should be 2. Threw exception TypeError: 'undefined' is not an object (evaluating 'subRule.operationType')
FAIL removeBaseURL(subRule.cssText) should be grayscale(1). Threw exception TypeError: 'undefined' is not an object (evaluating 'subRule.cssText')
FAIL subRule.operationType should be 12. Threw exception TypeError: 'undefined' is not an object (evaluating 'subRule.operationType')
FAIL removeBaseURL(subRule.cssText) should be custom(none url(fragment.shader), 1 1 filter-box). Threw exception TypeError: 'undefined' is not an object (evaluating 'subRule.cssText')
FAIL subRule.operationType should be 3. Threw exception TypeError: 'undefined' is not an object (evaluating 'subRule.operationType')
FAIL removeBaseURL(subRule.cssText) should be sepia(1). Threw exception TypeError: 'undefined' is not an object (evaluating 'subRule.cssText')
PASS successfullyParsed is true

TEST COMPLETE
